refactor(backend): drop the log redaction node budget
Rename the secrets helper to redaction. Remove the node counter that capped
how many containers one log line could expand. It truncated on traversal
order, so a later container lost data a reader needed.
